Fix bugs and restructures the frontend application
Fix bug on pagination when the user search a second time
on the same page.
Fix memory leaks due to fetch never close when a components
which calls getProductImage is unmount.
Fix wrong color on input-form (now use secondary color).

Change the functionnement of searchBar component to be reused
and more generic.

Create 3 mains categories for the applications (products,
recipes and stores).